Created successful training between environment, client, and server. Progress reports are bugged due to lack of episode termination

+++ RewardType.h
@@ -3,5 +3,5 @@ namespace RewardType
 {
 	const float PICKUP_BERRY = 1;
 
-	const float INVALID_ACTION = -0.05;
+	const float INVALID_ACTION = (float) -0.05;
 }
